API/requirements.txt
flask_restful
requests
flask
psycopg2-binary
flask
gunicorn
webargs
fuzzywuzzy
itsdangerous
python-Levenshtein
